chore: add config validation
Adds validation to ensure that a loaded config is always valid. For now
a config is valid if all fields are non-empty.

Apart from that there are no other requirements yet, but this may change
in the future if more configuration options are added.

Tests were added to test the validation logic. In addition a test case
was added to ensure that `config.sample.yaml` is always valid.