fix: do not publish docker images on PR
We don't really need this and it has one problem: we do not have any
garbage collections for PR images.

They should only be around as long as the PR is not merged, but this
would require some cleanup logic. Better not produce them in the first
place for now.