feat: add DaemonSet support to the helm chart
Adds the `kind` value which defaults to `Deployment` but can be set to
`DaemonSet` if desired.

Replica and HPA configuration is ignored if `kind` is `DaemonSet`.

Furthermore this change updates the `apiVersion` for the
`HorizontalPodAutoscaler` to `autoscaling/v2` since
`autoscaling/v2beta1` is deprecated and removed in Kubernetes 1.25. This
required some minor changes to the configuration schema.