fix: pass a `*runtime.Scheme` to the handler
The `admission.DecoderInjector` interface was removed in 0.15.0, so we
need to construct an `*admission.Decoder` ourselves from the runtime
scheme and pass it to the handler.

As of 0.15.0 there's a new "recommended" way to build a mutating webhook but it
requires bigger changes to the codebase that are better postponed until
after the dependency upgrade. Also it would break the API path of the
webhook since it'll get generated under a different name and there will
be no way to configure it.